Spawn now linked to data
Finalize linking spawn to data and add some events for handling load/process
NB may want to move the AssetBundle loading elsewhere and may want to have a loading screen while JSON data and AssetBundle loads (these take time) + Have the loading happen in background so the screen renders during load

+++ CAP6119Project-DataVisualization/Assets/Scripts/SpeciesManager.cs
@@ -13,7 +13,7 @@ public class SpeciesManager : MonoBehaviour
     public SpecimenDataManager DataMan;
     public string SpeciesName => species.name;
 
-    public float Distribution;
+    public double Distribution;
     public GameObject SpeciesPrefab;
 
     private List<GameObject> specimens; //instances of this species
@@ -48,7 +48,7 @@ public class SpeciesManager : MonoBehaviour
         SpeciesPrefab = model;
         model_lvl = lvl;
         
-        Distribution = species.count / TotalSampleCount;
+        Distribution = ((double)species.count) / (double)TotalSampleCount;
 
         _ready = true;
         spawned = false;
